Providers: Skip no further than the end of a buffer stream
skip() cut the length down to an Int before clamping it, so a large skip
wrapped around: Long.MAX_VALUE moved backwards and 2^32 + 2 skipped two.

+++ app/src/main/java/me/zhanghai/android/files/provider/common/ByteBufferInputStream.kt
@@ -36,7 +36,7 @@ class ByteBufferInputStream(buffer: ByteBuffer) : InputStream() {
         if (length <= 0) {
             return 0
         }
-        val skippedLength = length.toInt().coerceAtMost(buffer.remaining())
+        val skippedLength = length.coerceAtMost(buffer.remaining().toLong()).toInt()
         buffer.position(buffer.position() + skippedLength)
         return skippedLength.toLong()
     }
